Refactoring: Move the shared logic to AbstractUserCauseRestriction
