+++ src/main/java/com/synopsys/arc/jenkinsci/plugins/jobrestrictions/restrictions/job/StartedByMemberOfGroupRestriction.java
@@ -24,13 +24,7 @@
 package com.synopsys.arc.jenkinsci.plugins.jobrestrictions.restrictions.job;
 
 import hudson.Extension;
-import hudson.model.Action;
-import hudson.model.Cause;
-import hudson.model.CauseAction;
-import hudson.model.Queue;
-import hudson.model.Run;
 import hudson.model.User;
-import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -48,26 +42,21 @@ import com.synopsys.arc.jenkinsci.plugins.jobrestrictions.util.GroupSelector;
  * @author Oleg Nenashev <o.v.nenashev@gmail.com>
  * @since 0.4
  */
-public class StartedByMemberOfGroupRestriction extends JobRestriction {
+public class StartedByMemberOfGroupRestriction extends AbstractUserCauseRestriction {
 
     private final List<GroupSelector> groupList;
-    private final boolean checkGroupsFromUpstreamProjects;
 
     private transient Set<String> acceptedGroups = null;
 
     @DataBoundConstructor
     public StartedByMemberOfGroupRestriction(List<GroupSelector> groupList,
-            boolean checkGroupsFromUpstreamProjects) {
+            boolean checkUpstreamProjects) {
+        super(checkUpstreamProjects);
         this.groupList = groupList;
-        this.checkGroupsFromUpstreamProjects = checkGroupsFromUpstreamProjects;
     }
 
     public List<GroupSelector> getGroupList() {
         return groupList;
-    }
-
-    public boolean isCheckGroupsFromUpstreamProjects() {
-        return checkGroupsFromUpstreamProjects;
     }
 
     private synchronized @Nonnull Set<String> getAcceptedGroups() {
@@ -82,7 +71,8 @@ public class StartedByMemberOfGroupRestriction extends JobRestriction {
         return acceptedGroups;
     }
 
-    private boolean acceptsUser(@CheckForNull String userId) {
+    @Override
+    protected boolean acceptsUser(@CheckForNull String userId) {
         if (userId == null) {
             return false;
         }
@@ -100,68 +90,7 @@ public class StartedByMemberOfGroupRestriction extends JobRestriction {
         return false;
     }
 
-    /* package */ boolean canTake(@Nonnull List<Cause> causes) {
-        boolean userIdCause = false;
-        boolean rebuildCause = false;
-        boolean upstreamCause = false;
-        
-        boolean aUserIdWasNotAccepted = false;
-        boolean userIdCauseExists = false;
-        
-        for (@CheckForNull Cause cause : causes) { 
-            if (cause == null) {
-                continue; // Protection from the bug in old core versions
-            }
-                   
-            // Check user causes
-            if (cause.getClass().equals(Cause.UserIdCause.class) && !aUserIdWasNotAccepted) {
-                userIdCauseExists = true;
-                //if several userIdCauses exists, be defensive and don't allow if one is not accepted.
-                final @CheckForNull String startedBy = ((Cause.UserIdCause) cause).getUserId();
-                if (acceptsUser(startedBy)) {
-                    userIdCause = true;
-                } else {
-                    aUserIdWasNotAccepted = true;
-                    userIdCause = false;
-                }
-            }
-                      
-            // Check upstream projects if required
-            if (checkGroupsFromUpstreamProjects && cause.getClass().equals(Cause.UpstreamCause.class)) {
-                final List<Cause> upstreamCauses = ((Cause.UpstreamCause) cause).getUpstreamCauses();
-                // Recursive call to iterate through all above
-                if (canTake(upstreamCauses)) {
-                    upstreamCause = true;
-                }
-            }
-
-            // TODO: Check rebuild causes
-        }
-
-        //userId has preceedence
-        if (userIdCauseExists) {
-            return userIdCause;
-        } else { //If no update cause exists we should also return false...
-            return upstreamCause;
-        }
-    }
-
-    @Override
-    public boolean canTake(Queue.BuildableItem item) {
-        final List<Cause> causes = new ArrayList<Cause>();
-        for (Action action : item.getActions()) {
-            if (action instanceof CauseAction) {
-                CauseAction causeAction = (CauseAction) action;
-                causes.addAll(causeAction.getCauses());
-            } 
-        }
-        return canTake(causes);
-    }
-
-    @Override
-    public boolean canTake(Run run) {
-        return canTake(run.getCauses());
-    }
+    
 
     @Extension
     public static class DescriptorImpl extends JobRestrictionDescriptor {
