+++ src/main/resources/com/synopsys/arc/jenkinsci/plugins/jobrestrictions/restrictions/job/StartedByMemberOfGroupRestriction/config.properties
@@ -1,3 +1,3 @@
 groupList=Group list
-checkGroupsFromUpstreamProjects=Accept upstream runs started by groups above
+checkUpstreamProjects=Accept upstream runs started by groups above
