+++ src/main/resources/com/synopsys/arc/jenkinsci/plugins/jobrestrictions/restrictions/job/StartedByUserRestriction/config.jelly
@@ -29,8 +29,8 @@
     <f:entry title="${%usersList}"> 
         <f:repeatableProperty field="usersList" add="${%Add user}"/>  
     </f:entry>
-    <f:entry field="checkUsersFromUpstreamProjects">
-        <f:checkbox title="${%checkUsersFromUpstreamProjects}" checked="${it.checkUsersFromUpstreamProjects}"/>
+    <f:entry field="checkUpstreamProjects">
+        <f:checkbox title="${%checkUpstreamProjects}" checked="${it.checkUpstreamProjects}"/>
     </f:entry>
     <f:entry field="acceptAnonymousUsers">
         <f:checkbox title="${%acceptAnonymousUsers}" checked="${it.acceptAnonymousUsers}"/>
