+++ src/main/resources/com/synopsys/arc/jenkinsci/plugins/jobrestrictions/restrictions/job/StartedByUserRestriction/config.properties
@@ -1,3 +1,3 @@
 usersList=Users list
-checkUsersFromUpstreamProjects=Accept upstream runs started by users above
+checkUpstreamProjects=Accept upstream runs started by users above
 acceptAnonymousUsers=Accept runs started by anonymous users
